move some functions from helper
